Fix cast

+++ mock/src/generator.cpp
@@ -63,7 +63,7 @@ void Generator::MainLoop()
         if (overrange)
             record->header->record_status |= ADQ_RECORD_STATUS_OVERRANGE;
 
-        record->header->timestamp_synchronization_counter = record_number / 100;
+        record->header->timestamp_synchronization_counter = static_cast<uint16_t>(record_number / 100);
         if (!(record_number % 50))
             record->header->misc |= 0x1u;
         if (!(record_number % 30))
